# frontend/默认模块.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/articles/{id}/publish": {
      "post": {
        "summary": "publish",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/admin/users": {
      "get": {
        "summary": "page",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "pageNo",
            "in": "query",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIPageAdminUserVO"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": {
                    "records": [
                      {
                        "id": 0,
                        "email": "",
                        "nickname": "",
                        "role": "",
                        "status": "",
                        "createTime": "",
                        "lastLoginTime": ""
                      }
                    ],
                    "total": 0,
                    "size": 0,
                    "current": 0,
                    "orders": [
                      {
                        "column": "",
                        "asc": false
                      }
                    ],
                    "optimizeCountSql": false,
                    "searchCount": false,
                    "optimizeJoinOfCountSql": false,
                    "maxLimit": 0,
                    "countId": "",
                    "pages": 0
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/users/{id}/status": {
      "post": {
        "summary": "updateStatus",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": false
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/users/{id}/recharge": {
      "post": {
        "summary": "recharge",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "amount",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": false
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/admin/articles": {
      "get": {
        "summary": "管理后台 - 文章分页列表",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "pageNo",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIPageArticleListVO"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": {
                    "records": [
                      {
                        "id": 0,
                        "title": "",
                        "summary": "",
                        "coverUrl": "",
                        "type": "",
                        "price": 0,
                        "status": "",
                        "publishTime": "",
                        "viewCount": 0,
                        "likeCount": 0,
                        "pinned": false,
                        "pinnedTime": ""
                      }
                    ],
                    "total": 0,
                    "size": 0,
                    "current": 0,
                    "orders": [
                      {
                        "column": "",
                        "asc": false
                      }
                    ],
                    "optimizeCountSql": false,
                    "searchCount": false,
                    "optimizeJoinOfCountSql": false,
                    "maxLimit": 0,
                    "countId": "",
                    "pages": 0
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/articles/{id}/publish": {
      "post": {
        "summary": "publish",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/articles/{id}/offline": {
      "post": {
        "summary": "offline",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/articles/{id}/pin": {
      "post": {
        "summary": "pin",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/articles/{id}/unpin": {
      "post": {
        "summary": "unpin",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "description": "具体数据内容",
            "type": "null"
          }
        }
      },
...
      "ArticleListVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "description": ""
          },
          "summary": {
            "type": "string",
            "description": ""
          },
          "coverUrl": {
            "type": "string",
            "description": ""
          },
          "type": {
            "type": "string",
            "description": ""
          },
          "price": {
            "type": "number",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": ""
          },
          "publishTime": {
            "type": "string",
            "description": ""
          },
          "viewCount": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "likeCount": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "pinned": {
            "type": "boolean",
            "description": ""
          },
          "pinnedTime": {
            "type": "string",
            "description": ""
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "column": {
            "type": "string",
            "description": ""
          },
          "asc": {
            "type": "boolean",
            "description": ""
          }
        }
      },
      "IPageArticleListVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleListVO",
              "description": "文章列表项。"
            },
            "description": ""
          },
          "total": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem",
              "description": "com.baomidou.mybatisplus.core.metadata.OrderItem"
            },
            "description": ""
          },
          "optimizeCountSql": {
            "type": "boolean",
            "description": ""
          },
          "searchCount": {
            "type": "boolean",
            "description": ""
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "description": ""
          },
          "maxLimit": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "countId": {
            "type": "string",
            "description": ""
          },
          "pages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ApiResponseIPageArticleListVO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "$ref": "#/components/schemas/IPageArticleListVO",
            "description": "具体数据内容"
          }
        }
      },
...
      "AdminUserVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "email": {
            "type": "string",
            "description": ""
          },
          "nickname": {
            "type": "string",
            "description": ""
          },
          "role": {
            "type": "string",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": ""
          },
          "createTime": {
            "type": "string",
            "description": ""
          },
          "lastLoginTime": {
            "type": "string",
            "description": ""
          },
          "balance": {
            "type": "number",
            "description": ""
          }
        }
      },
      "IPageAdminUserVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserVO",
              "description": "com.bluemsun.blog.module.admin.vo.AdminUserVO"
            },
            "description": ""
          },
          "total": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem",
              "description": "com.baomidou.mybatisplus.core.metadata.OrderItem"
            },
            "description": ""
          },
          "optimizeCountSql": {
            "type": "boolean",
            "description": ""
          },
          "searchCount": {
            "type": "boolean",
            "description": ""
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "description": ""
          },
          "maxLimit": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "countId": {
            "type": "string",
            "description": ""
          },
          "pages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ApiResponseIPageAdminUserVO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "$ref": "#/components/schemas/IPageAdminUserVO",
            "description": "具体数据内容"
          }
        }
      },
      "ApiResponseBoolean": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "type": "boolean",
            "description": "具体数据内容"
          }
        }
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}